backend/requirement.txt
fastapi==0.95.0
uvicorn==0.21.1
websockets==10.4
aiortc==1.4.0
pydantic==1.10.7
python-multipart==0.0.6
uuid==1.30
python-dotenv==1.0.0